Refactor

+++ Core/Categories/AutoCategory.cs
@@ -2,7 +2,7 @@
 {
     public class AutoCategory : Category
     {
-        public AutoCategory(string name, double increment, double capacity, string category)
+        public AutoCategory(string name, int increment, int capacity, string category)
              : base(name, increment, capacity)
         {
             Category = category;
