+++ Core/Categories/RegexCategory.cs
@@ -5,7 +5,7 @@ namespace Core
 {
     public class RegexCategory : Category
     {
-        public RegexCategory(string name, IReadOnlyCollection<Rule> rules, double increment, double capacity)
+        public RegexCategory(string name, IReadOnlyCollection<Rule> rules, int increment, int capacity)
             :base(name, increment, capacity)
         {
             Rules = rules;
